Rename to FpsCounter
