+++ src/Engine/UI/Screen.java
@@ -1,7 +1,6 @@
 package Engine.UI;
 
-import Engine.Core.FpsMeasureThread;
-import Engine.Core.KeyBinding;
+import Engine.FpsMeasure;
 import Engine.Helper.RenderHelper;
 import Engine.RenderSetting;
 
@@ -16,7 +15,7 @@ public abstract class Screen extends JPanel {
     /**
      * The FPS counter thread.
      */
-    protected final FpsMeasureThread fpsMeasureThread;
+    protected final FpsMeasure fpsMeasure;
     /**
      * The thread in charge of the rendering.
      */
@@ -44,7 +43,7 @@ public abstract class Screen extends JPanel {
 
         this.onDemandRender = false;
         this.targetFps = targetFps;
-        this.fpsMeasureThread = new FpsMeasureThread();
+        this.fpsMeasure = new FpsMeasure();
         this.renderThread = new Thread(() -> {
             while (true) {
                 this.repaint();
@@ -57,23 +56,6 @@ public abstract class Screen extends JPanel {
      */
     public Screen() {
         this(RenderSetting.maxFps);
-    }
-
-    /**
-     * Register a key binding.
-     *
-     * @param bind The key binding.
-     */
-    public void registerKeyEvent(KeyBinding bind) {
-        this.getInputMap(JComponent.WHEN_IN_FOCUSED_WINDOW).put(
-                KeyStroke.getKeyStroke(bind.getKeys()),
-                bind.getNote()
-        );
-
-        this.getActionMap().put(
-                bind.getNote(),
-                bind.getAction()
-        );
     }
 
     /**
@@ -134,7 +116,7 @@ public abstract class Screen extends JPanel {
 
         try {
             render(g2d);
-            this.fpsMeasureThread.interrupt();
+            this.fpsMeasure.interrupt();
             Thread.sleep((long) RenderHelper.getRenderDelayForTargetFps(this.targetFps));
         } catch (InterruptedException e) {
             throw new RuntimeException(e);
@@ -153,7 +135,7 @@ public abstract class Screen extends JPanel {
         try {
             if (!this.onDemandRender) {
                 this.renderThread.start();
-                this.fpsMeasureThread.start();
+                this.fpsMeasure.start();
             }
         } catch (IllegalThreadStateException ex) {
             // fuck you.
@@ -165,7 +147,7 @@ public abstract class Screen extends JPanel {
      */
     public void dispose() {
         if (this.renderThread.isAlive()) this.renderThread.interrupt();
-        if (this.fpsMeasureThread.isAlive()) this.fpsMeasureThread.interrupt();
+        if (this.fpsMeasure.isAlive()) this.fpsMeasure.interrupt();
     }
 
     /**
